requirements.txt
discord.py==2.3.2
selenium==4.20.0
